
\docType{data}
\name{Kernel Look Up Tables}
\alias{simulated_CV}
\title{Fixed Critical Values using the Simulated Method}
\format{
\itemize{
  \item d: dimensions
  \item alpha: objective type 1 error rate
}
}
\usage{
Bartlett_Mother_01_Master
}
\description{
A look up table containing the simulated robust critical values.
}

\keyword{datasets}

\details{
The title of each data set is the <Mother>_<Lugsail>_<alpha>_Master
}
