% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_CV.R
\docType{data}
\name{fitted_CV}
\alias{fitted_CV}
\title{Fixed Critical Values using the Fitted Method}
\format{
A data frame with the following columns:
\itemize{
\item intercept:a parameter for the approximation
\item beta1: a parameter for the approximation
\item beta2: a parameter for the approximation
\item beta3: a parameter for the approximation
\item adj.r.sq: Adjusted \eqn{R^2}
\item d: dimensions
\item kernel: mother kernel selected
\item lugsail: lugsail setting
\item alpha: objective type 1 error rate
}
}
\usage{
fitted_CV
}
\description{
A look up table containing the fitted coefficients for robust critical values. It also contains the adjusted \eqn{R^2} values for the respective fit.
}
\details{
This data set contains critical values that the \code{get_cv} and \code{get_cv_fitted} functions search for.
For a given kernel, lugsail setting, dimension, and desired \eqn{\alpha}, the critical value is approximated via the formula,
\eqn{CV = \text{intercept} + \beta_1*b+ \beta_2*b^2 + \beta_3*b^3}
}
\keyword{datasets}
