% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cv.R
\name{generate_cv}
\alias{generate_cv}
\title{Generate Fixed-b Critical Values}
\usage{
generate_cv(
  b,
  d = 2,
  alpha = 0.05,
  the_kernel = bartlett,
  lugsail_type = "Mother",
  q = 1,
  return_F_stats = F,
  num_replicates = 50000,
  replicate_size = 1000
)
}
\arguments{
\item{b}{Numeric vector of bandwidth values between 0 and 1. The bandwidth
represents the proportion of autocovariances given non-zero weight.}

\item{d}{Integer, the dimension (degrees of freedom) of the test statistic.
Default is 1.}

\item{alpha}{Numeric vector of desired significance levels between 0 and 1.
Default is 0.05.}

\item{the_kernel}{A function for the kernel used to generate the weights. The
package supports Bartlett (\code{bartlett}), Parzen (\code{parzen}),
Tukey-Hanning (\code{th}), or Quadratic Spectral \code{qs}.
User supplied functions are also supported.}

\item{lugsail_type}{Character string specifying the lugsail transformation. Options are
\code{"Mother"} (default), \code{"Zero"}, or \code{"Over"}.}

\item{q}{The Parzen characteristic exponent of the kernel supplied. The default is 1.}

\item{return_F_stats}{A logical value indicating if test statistics used to
approximate the fixed-b distribution should be returned. The default value is FALSE.}

\item{num_replicates}{A integer indicating of test statistics to be generated used to
approximate the fixed-b distribution. Default is 50,000.}

\item{replicate_size}{A integer indicating the sample size used to construct
the test statistics that are used to approximate the fixed-b distribution. Default is 1000.}
}
\value{
If \code{return_F_stats = FALSE} (default) a matrix is returned where the rows
correspond to the values for \code{b} and the columns correspond to the
significance levels (\code{alpha}) supplied by the user. If \code{return_F_stats = TRUE},
then a list is returned. The second element in the list is a matrix of the F
statistics generated to approximate the fixed-b distribution.
}
\description{
Generates fixed-b critical values for robust hypothesis testing. Simulate from
a fixed-b distribution to get custom critical values.
}
\details{
This function is used to generate custom simulated critical values.
Both package kernels and user supplied kernels maybe supplied. It is not recommended to decrease
 \code{num_replicates} or \code{replicate_size}, especially for higher dimensions.
}
\examples{
# Multiple b's and alpha's
set.seed(62)
CVs <- generate_cv(b = c(0.005, 0.006), d=1, alpha = c(0.05, 0.01),
num_replicates = 100)
CVs
# Custom kernel function
my_kernel <- function(x){
  k_x <- exp(-x^2/2)
  return(k_x)
}
CVs <- generate_cv(b = 0.005, d=1, alpha = 0.05, the_kernel = my_kernel,
 q = 2, num_replicates = 100)
CVs
}
