% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GillStream}
\alias{GillStream}
\title{Stream Temperature data}
\format{
A data frame with 153 rows and 3 columns:
\describe{
  \item{Date}{The date the temperature was recorded.}
  \item{AvgTemp}{Temperature in Celsius}
  \item{time_index}{index values for the date}
}
}
\usage{
GillStream
}
\description{
This data set contains daily average temperature readings in Celsius for a
stream in Gill, Massachusetts. The original data was recorded via
a submerged logger at 15 minutes intervals from August to December
2024 by the Organismic & Evolutionary Biology Department at the
University of Massachusetts Amherst.
}
\keyword{datasets}
