% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_CV}
\alias{simulated_CV}
\alias{Bartlett_Mother_01_Master}
\alias{Bartlett_Mother_025_Master}
\alias{Bartlett_Mother_05_Master}
\alias{Bartlett_Mother_10_Master}
\alias{Bartlett_Zero_01_Master}
\alias{Bartlett_Zero_025_Master}
\alias{Bartlett_Zero_05_Master}
\alias{Bartlett_Zero_10_Master}
\alias{Bartlett_Over_01_Master}
\alias{Bartlett_Over_025_Master}
\alias{Bartlett_Over_05_Master}
\alias{Bartlett_Over_10_Master}
\alias{QS_Mother_01_Master}
\alias{QS_Mother_025_Master}
\alias{QS_Mother_05_Master}
\alias{QS_Mother_10_Master}
\alias{QS_Zero_01_Master}
\alias{QS_Zero_025_Master}
\alias{QS_Zero_05_Master}
\alias{QS_Zero_10_Master}
\title{Fixed Critical Values using the Simulated Method}
\format{
A look up table containing the simulated robust critical values. The 'b' column contains the bandwidth, and subsequent columns are the dimensions
\itemize{
  \item b: bandwidth of the test statistics. The proportion of autocovariances given a non-zero weight.
  \item d: the dimension of the test statistic.
}
}
\usage{
Bartlett_Mother_01_Master

Bartlett_Mother_025_Master

Bartlett_Mother_05_Master

Bartlett_Mother_10_Master

Bartlett_Zero_01_Master

Bartlett_Zero_025_Master

Bartlett_Zero_05_Master

Bartlett_Zero_10_Master

Bartlett_Over_01_Master

Bartlett_Over_025_Master

Bartlett_Over_05_Master

Bartlett_Over_10_Master

QS_Mother_01_Master

QS_Mother_025_Master

QS_Mother_05_Master

QS_Mother_10_Master

QS_Zero_01_Master

QS_Zero_025_Master

QS_Zero_05_Master

QS_Zero_10_Master
}
\description{
The title of each data set is the \code{<Mother>_<Lugsail>_<alpha>_Master}.
}
\details{
The title of each data set is the \code{<Mother>_<Lugsail>_<alpha>_Master}. Supported mother kernels are Bartlett, Parzen, Tukey-Hanning (TH), and Quadratic Spectral (QS). Supported lugsail settings are mother, zero, and over. The available critical values are 0.01, 0.025, 0.05, and 0.10.
Note that when b = 0 the robust critical value are equivalent to chi-square critical values.
}
\keyword{datasets}
